Clean up obsolete files from persistence with an E, move the useful ones to persistance with an A, rename files that had to be renamed, give more descriptive names and rename imports in many parts of the file.
