--- modules/fidesModule/persistence/__init__.py
@@ -1,1 +1,0 @@
-# classes used to access persistence in as persistent storage
